RF: Move `disable_in_read_only_mode()` into `api` package
The function is only used within the `api` package
