+++ datalad_registry/utils/flask_tools.py
@@ -1,9 +1,4 @@
-from functools import wraps
-from http import HTTPStatus
-
-from flask import Response, abort, current_app
-
-from datalad_registry.conf import OperationMode
+from flask import Response, current_app
 
 
 def json_resp_from_str(json_str: str, *args, **kwargs) -> Response:
@@ -22,27 +17,3 @@ def json_resp_from_str(json_str: str, *args, **kwargs) -> Response:
     return current_app.response_class(
         json_str, *args, mimetype="application/json", **kwargs
     )
-
-
-def disable_in_read_only_mode(view_func):
-    """
-    Decorator for view functions that should be disabled in read-only mode
-
-    :param view_func: The view function to decorate
-    :return: The decorated view function
-    """
-
-    @wraps(view_func)
-    def wrapper(*args, **kwargs):
-        if (
-            current_app.config["DATALAD_REGISTRY_OPERATION_MODE"]
-            is OperationMode.READ_ONLY
-        ):
-            abort(
-                HTTPStatus.IM_A_TEAPOT,
-                description="This operation is not available from a read-only server.",
-            )
-        else:
-            return view_func(*args, **kwargs)
-
-    return wrapper
